treebeard

+++ src/collaborative_platform/assets/js/scripts/projectFileTree.js
@@ -5,9 +5,12 @@ $.ajaxSetup({
     }
 });
 
+var id = $('#files').attr('data-project-id');
+var draggableElements;
+
 var options = {
     divID : 'files',
-    filesData : "/static/tree.json",
+    filesData : '/api/files/get_tree/' + id,
     rowHeight : 35,
     showTotal : 15,
     paginate : false,
@@ -28,7 +31,7 @@ var options = {
                 filter : true,
                 css : 'tb-draggable',
                 custom : function (row) {
-                    return m("a[href='asdasd']", row.data.title)
+                    return  (row.data.kind !== "folder") ? m("a[href='/projects/']", row.data.name) : m("span", row.data.name)
                 }
             },
             {
@@ -49,17 +52,28 @@ var options = {
                     var editButton = m("button.tb-button", {
                             onclick: function _editClick(e) {
                                 e.stopPropagation();
-                                var value = row.data.title;
+                                var title = row.data.name;
+                                var value = '';
+
+
+                                var data = "d"
+                                function update(e) {
+                                    if (data.length < 5) data = e.target.value
+                                }
+
+
                                 var mithrilContent = m('div', [
                                     m('h3.break-word', 'Change name'),
-                                    m('p', 'Name:'),
-                                    m('input', {class: 'form-control', value: value, oninput: function (e) {value = e.target.value}})
+                                    m('p', 'Old name: ' + title),
+                                    m('p', 'New name:'),
+                                    m('input.form-control', {oninput: function (e) {value = e.target.value}}  )
                                 ]);
+
                                 var mithrilButtons = m('div', [
                                     m('button', { 'class' : 'btn btn-default m-r-md', onclick : function() { that.modal.dismiss(); } }, 'Cancel'),
                                     m('button', { 'class' : 'btn btn-success', onclick : function() {
 
-                                        var url = (row.data.kind !== "folder") ? '/api/files/rename/file/' : '/api/files/rename/directory/' + row.data.id + '/';
+                                        var url = (row.data.kind !== "folder") ? '/api/files/' + row.data.id + '/rename/' : '/api/files/directory/' + row.data.id + '/rename/';
 
                                         $.ajax({
                                             type: "PUT",
@@ -84,7 +98,7 @@ var options = {
                             var mithrilContent = m('div', [
                                 m('h3.break-word', 'Create folder'),
                                 m('p', 'Name:'),
-                                m('input', {class: 'form-control', value: value, oninput: function (e) {value = e.target.value}})
+                                m('input', {class: 'form-control', oninput: function (e) {value = e.target.value}} )
                             ]);
                             var mithrilButtons = m('div', [
                                 m('button', { 'class' : 'btn btn-default m-r-md', onclick : function() { that.modal.dismiss(); } }, 'Cancel'),
@@ -92,7 +106,7 @@ var options = {
 
                                     $.ajax({
                                         type: "PUT",
-                                        url: '/api/files/create/directory/' + row.data.id + '/' + value,
+                                        url: '/api/files/directory/' + row.data.id + '/create_subdir/' + value,
                                         contentType : 'application/json',
                                         success: function(resultData){
                                             that.modal.dismiss();
@@ -110,13 +124,13 @@ var options = {
                         onclick: function _deleteClick(e) {
                             e.stopPropagation();
                             var mithrilContent = m('div', [
-                                m('h3.break-word', 'Delete "' + row.data.title+ '"?'),
+                                m('h3.break-word', 'Delete "' + row.data.name + '"?'),
                                 m('p', 'This action is irreversible.')
                             ]);
                             var mithrilButtons = m('div', [
                                 m('button', { 'class' : 'btn btn-default m-r-md', onclick : function() { that.modal.dismiss(); } }, 'Cancel'),
                                 m('button', { 'class' : 'btn btn-success', onclick : function() {
-                                    var url = (row.data.kind !== "folder") ? '/api/files/file/' : '/api/files/directory/';
+                                    var url = (row.data.kind !== "folder") ? '/api/files/' : '/api/files/directory/';
 
                                     $.ajax({
                                         type: "DELETE",
@@ -134,17 +148,33 @@ var options = {
                         }
                     }, m("i", {'class': "fa fa-trash"}));
 
-                    return (row.data.kind !== "folder") ? [editButton, deleteButton] : [editButton, createFolderButton, deleteButton];
+                    if (row.data.parent) {
+                        return (row.data.kind !== "folder") ? [editButton, deleteButton] : [editButton, createFolderButton, deleteButton];
+                    } else {
+                        return createFolderButton;
+                    }
+
+
                 }
             }
         ];
     },
+    dragEvents : {
+        start : function(event, ui) {
+            console.log('start')
+            $('.tb-multiselect').each(function(index) {
+                console.log(this)
+            })
+        },
+    },
     dropEvents : {
+
         over : function(event, ui) {
+            //console.log(ui)
         },
         drop : function(event, ui) {
-            console.log(ui)
-            console.log(event)
+            //console.log(ui)
+            //console.log(event)
             //var obj = ui.draggable
             //console.log(obj)
